move sort to discussions state
